feat: skip RK modal if we know is CB browser
