+++ src/components/ConnectDialog/ConnectDialog.tsx
@@ -2,6 +2,8 @@ import { FC, useEffect } from 'react'
 import { ConnectButton } from '@rainbow-me/rainbowkit'
 import clsx from 'clsx'
 import { Button, ButtonProps } from '../Button'
+import { useConnectCBWallet } from '@/utils/useConnectCBWallet'
+import { getIsCoinbaseBrowser } from '@/utils/getIsCoinbaseBrowser'
 
 type ConnectDialogProps = {
   title?: React.ReactNode
@@ -14,6 +16,8 @@ export const ConnectDialog: FC<ConnectDialogProps> = ({
   inNavbar = false,
   size,
 }) => {
+  const connect = useConnectCBWallet()
+
   return (
     <ConnectButton.Custom>
       {({
@@ -47,7 +51,13 @@ export const ConnectDialog: FC<ConnectDialogProps> = ({
               if (!connected) {
                 return (
                   <Button
-                    onClick={openConnectModal}
+                    onClick={() => {
+                      if (getIsCoinbaseBrowser()) {
+                        connect()
+                      } else {
+                        openConnectModal()
+                      }
+                    }}
                     type="button"
                     className={clsx({
                       'rounded-lg !py-1.5 !px-4': inNavbar,
